fix(synthetic-shadow): cleaning up tests (#1334)
* fix(engine): issue #1311 to remove assertion mock

* fix(synthetic-shadow): move activeElement test to karma

* Update packages/integration-karma/test/polyfills/shadow-root/index.spec.js


* Update packages/integration-karma/test/polyfills/shadow-root/index.spec.js


* Update packages/integration-karma/test/polyfills/shadow-root/index.spec.js


* fix(integration-karma): types
