+++ packages/@lwc/engine/src/framework/__tests__/html-element.spec.ts
@@ -1467,77 +1467,4 @@ describe('html-element', () => {
             });
         });
     });
-
-    describe('.template', () => {
-        describe('.activeElement', () => {
-            it('should be null when no active element is found', () => {
-                let template;
-                const html = compileTemplate(`
-                    <template>
-                        <input />
-                    </template>
-                `);
-                class Foo extends LightningElement {
-                    constructor() {
-                        super();
-                        template = this.template;
-                    }
-                    render() {
-                        return html;
-                    }
-                }
-                const elm = createElement('x-parent', { is: Foo });
-                document.body.appendChild(elm);
-                expect(template.activeElement).toBe(null);
-            });
-
-            it('should be null when the active element is outside of the shadow', () => {
-                let template;
-                const html = compileTemplate(`
-                    <template>
-                        <input />
-                    </template>
-                `);
-                class Foo extends LightningElement {
-                    constructor() {
-                        super();
-                        template = this.template;
-                    }
-                    render() {
-                        return html;
-                    }
-                }
-                const elm = createElement('x-parent', { is: Foo });
-                const outsideInput = document.createElement('input');
-                document.body.appendChild(elm);
-                document.body.appendChild(outsideInput);
-                outsideInput.focus();
-                expect(template.activeElement).toBe(null);
-                expect(document.activeElement).toBe(outsideInput);
-            });
-
-            it('should be a local input when focused', () => {
-                let template;
-                const html = compileTemplate(`
-                    <template>
-                        <input />
-                    </template>
-                `);
-                class Foo extends LightningElement {
-                    constructor() {
-                        super();
-                        template = this.template;
-                    }
-                    render() {
-                        return html;
-                    }
-                }
-                const elm = createElement('x-parent', { is: Foo });
-                document.body.appendChild(elm);
-                const input = template.querySelector('input');
-                input.focus();
-                expect(template.activeElement).toBe(input);
-            });
-        });
-    });
 });
